+++ hutool-core/src/main/java/cn/hutool/core/util/NumberUtil.java
@@ -2044,6 +2044,11 @@ public class NumberUtil {
 	 * @since 4.0.9
 	 */
 	public static BigDecimal toBigDecimal(String number) {
+		try{
+			number = parseNumber(number).toString();
+		} catch (Exception ignore){
+			// 忽略解析错误
+		}
 		return StrUtil.isBlank(number) ? BigDecimal.ZERO : new BigDecimal(number);
 	}
 
@@ -2287,7 +2292,7 @@ public class NumberUtil {
 	 *
 	 * <pre>
 	 * 1、0x开头的视为16进制数字
-	 * 2、0开头的视为8进制数字
+	 * 2、0开头的忽略开头的0
 	 * 3、其它情况按照10进制转换
 	 * 4、空串返回0
 	 * 5、.123形式返回0（按照小于0的小数对待）
@@ -2304,18 +2309,16 @@ public class NumberUtil {
 			return 0;
 		}
 
-		// 对于带小数转换为整数采取去掉小数的策略
-		number = StrUtil.subBefore(number, CharUtil.DOT, false);
-		if (StrUtil.isEmpty(number)) {
-			return 0;
-		}
-
 		if (StrUtil.startWithIgnoreCase(number, "0x")) {
 			// 0x04表示16进制数
 			return Integer.parseInt(number.substring(2), 16);
 		}
 
-		return Integer.parseInt(removeNumberFlag(number));
+		try{
+			return Integer.parseInt(number);
+		} catch (NumberFormatException e){
+			return parseNumber(number).intValue();
+		}
 	}
 
 	/**
@@ -2323,9 +2326,11 @@ public class NumberUtil {
 	 *
 	 * <pre>
 	 * 1、0x开头的视为16进制数字
-	 * 2、0开头的视为8进制数字
+	 * 2、0开头的忽略开头的0
 	 * 3、空串返回0
 	 * 4、其它情况按照10进制转换
+	 * 5、.123形式返回0（按照小于0的小数对待）
+	 * 6、123.56截取小数点之前的数字，忽略小数部分
 	 * </pre>
 	 *
 	 * @param number 数字，支持0x开头、0开头和普通十进制
@@ -2334,13 +2339,7 @@ public class NumberUtil {
 	 */
 	public static long parseLong(String number) {
 		if (StrUtil.isBlank(number)) {
-			return 0;
-		}
-
-		// 对于带小数转换为整数采取去掉小数的策略
-		number = StrUtil.subBefore(number, CharUtil.DOT, false);
-		if (StrUtil.isEmpty(number)) {
-			return 0;
+			return 0L;
 		}
 
 		if (number.startsWith("0x")) {
@@ -2348,7 +2347,63 @@ public class NumberUtil {
 			return Long.parseLong(number.substring(2), 16);
 		}
 
-		return Long.parseLong(removeNumberFlag(number));
+		try{
+			return Long.parseLong(number);
+		} catch (NumberFormatException e){
+			return parseNumber(number).longValue();
+		}
+	}
+
+	/**
+	 * 解析转换数字字符串为long型数字，规则如下：
+	 *
+	 * <pre>
+	 * 1、0开头的忽略开头的0
+	 * 2、空串返回0
+	 * 3、其它情况按照10进制转换
+	 * 4、.123形式返回0.123（按照小于0的小数对待）
+	 * </pre>
+	 *
+	 * @param number 数字，支持0x开头、0开头和普通十进制
+	 * @return long
+	 * @since 5.5.5
+	 */
+	public static float parseFloat(String number) {
+		if (StrUtil.isBlank(number)) {
+			return 0f;
+		}
+
+		try{
+			return Float.parseFloat(number);
+		} catch (NumberFormatException e){
+			return parseNumber(number).floatValue();
+		}
+	}
+
+	/**
+	 * 解析转换数字字符串为long型数字，规则如下：
+	 *
+	 * <pre>
+	 * 1、0开头的忽略开头的0
+	 * 2、空串返回0
+	 * 3、其它情况按照10进制转换
+	 * 4、.123形式返回0.123（按照小于0的小数对待）
+	 * </pre>
+	 *
+	 * @param number 数字，支持0x开头、0开头和普通十进制
+	 * @return long
+	 * @since 5.5.5
+	 */
+	public static double parseDouble(String number) {
+		if (StrUtil.isBlank(number)) {
+			return 0D;
+		}
+
+		try{
+			return Double.parseDouble(number);
+		} catch (NumberFormatException e){
+			return parseNumber(number).doubleValue();
+		}
 	}
 
 	/**
@@ -2357,13 +2412,15 @@ public class NumberUtil {
 	 * @param numberStr Number字符串
 	 * @return Number对象
 	 * @since 4.1.15
+	 * @throws NumberFormatException 包装了{@link ParseException}，当给定的数字字符串无法解析时抛出
 	 */
-	public static Number parseNumber(String numberStr) {
-		numberStr = removeNumberFlag(numberStr);
+	public static Number parseNumber(String numberStr) throws NumberFormatException{
 		try {
 			return NumberFormat.getInstance().parse(numberStr);
 		} catch (ParseException e) {
-			throw new UtilException(e);
+			final NumberFormatException nfe = new NumberFormatException(e.getMessage());
+			nfe.initCause(e);
+			throw nfe;
 		}
 	}
 
@@ -2508,26 +2565,6 @@ public class NumberUtil {
 		} else {
 			return selectNum * mathNode(selectNum - 1);
 		}
-	}
-
-	/**
-	 * 去掉数字尾部的数字标识，例如12D，44.0F，22L中的最后一个字母
-	 *
-	 * @param number 数字字符串
-	 * @return 去掉标识的字符串
-	 */
-	private static String removeNumberFlag(String number) {
-		// 去掉千位分隔符
-		if (StrUtil.contains(number, CharUtil.COMMA)) {
-			number = StrUtil.removeAll(number, CharUtil.COMMA);
-		}
-		// 去掉类型标识的结尾
-		final int lastPos = number.length() - 1;
-		final char lastCharUpper = Character.toUpperCase(number.charAt(lastPos));
-		if ('D' == lastCharUpper || 'L' == lastCharUpper || 'F' == lastCharUpper) {
-			number = StrUtil.subPre(number, lastPos);
-		}
-		return number;
 	}
 	// ------------------------------------------------------------------------------------------- Private method end
 }